trip 앱 생성
