+++ photos/serializers.py
@@ -2,12 +2,6 @@ from rest_framework import serializers
 from .models import *
 
 
-class TripSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Trip
-        exclude = ['created_at', 'updated_at']
-
-
 class TagYoloSerializer(serializers.ModelSerializer):
     class Meta:
         model = TagYolo
