+++ tripfriend/urls.py
@@ -18,5 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('api.urls')),
+    path('accounts/', include('accounts.urls')),
+    path('photos/', include('photos.urls')),
 ]
